PyCharmMiscProject/requirements.txt.txt
kivy==2.3.1
kivymd==1.1.1
plyer
gspread
google-auth